requirements.txt
openai-whisper>=20231117
# Pin Triton < 3 to avoid JITCallable._set_src() API break with word_timestamps
triton>=2.0.0,<3.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
